requirements-docs.txt
mkdocs==1.6.1
mkdocs-material==9.7.4
pymdown-extensions
mike
PyYAML

